Initial refactor of the tracing client.
Responsabilities are now separated into different classes. The Tracer
class contains the main loop which iterates over the different states
the execution goes through, as well as some additional utility functions
to aid with the tracing of a process.

The Preamble class contains all the logic to detect when a new trace
should begin. The RegionOfInterest class contains the logic to handle
what happens during the execution of a trace. Finally, the Prologue
contains the logic necessary to finish a trace and preparing the
system for the next trace.