Add argument passing to dynamic calls
This commit allows for arguments to be sent to the LD_PRELOAD'ed
library during tracing. To use this feature, fill the vector passed to dyn_call
with the necessary arguments. They should be unsigned longs to ensure maximum
compatibility across architectures. The user should keep these arguments to the
minimum since each platform will have a limited amount of argument "slots"
implemented.

In order to implement such features for an architecture, the "set_args" function
must be implemented. For reference on how to do so, you should check the
architecture's and linux's ABI and C call conventions. In this commit, the
implementation for the RiscV architecture is supplied, which could serve as
example for an architecture which has register argument passing.